+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.0.6 (upcoming)
 
 - Dropped ext-mbstring requirement
+- Update elixir manifest caching
 
 ## 4.0.5 (released 2016-03-20)